Update all lobbies

+++ src/controllers/theme_controller.cr
@@ -46,10 +46,18 @@ class ThemeController < ApplicationController
   end
 
   def destroy
-    theme.destroy
-    Lobby.where(theme_id: theme.id).select.each { |l| l.update(active: false) }
-    respond_with(204) do
-      json ""
+    lobbies = Lobby.where(theme_id: theme.id).select
+    running_games = Game.where(lobby_id: lobbies.map(&.id!), running: true).select
+    if running_games.empty?
+      lobbies.each { |l| l.update(active: false) }
+      theme.destroy
+      respond_with(204) do
+        json ""
+      end
+    else
+      respond_with(403) do
+        json({errors: [{"games": "A game is still running"}]}.to_json)
+      end
     end
   end
